Moved starting the details activity to the host activiy
