+++ ginivision/src/main/java/net/gini/android/vision/returnassistant/ReturnAssistantFragment.kt
@@ -2,7 +2,6 @@ package net.gini.android.vision.returnassistant
 
 import android.annotation.SuppressLint
 import android.app.Activity
-import android.content.Intent
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.support.v7.widget.LinearLayoutManager
@@ -127,8 +126,8 @@ class ReturnAssistantFragment : Fragment(), ReturnAssistantScreenContract.View,
         presenter?.deselectLineItem(lineItem)
     }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        presenter?.onActivityResult(requestCode, resultCode, data)
+    override fun updateLineItem(selectableLineItem: SelectableLineItem) {
+        presenter?.updateLineItem(selectableLineItem)
     }
+
 }
