+++ ginivision/src/main/java/net/gini/android/vision/returnassistant/ReturnAssistantFragmentInterface.kt
@@ -9,4 +9,5 @@ interface ReturnAssistantFragmentInterface {
 
     var listener: ReturnAssistantFragmentListener?
 
+    fun updateLineItem(selectableLineItem: SelectableLineItem)
 }
