+++ ginivision/src/main/java/net/gini/android/vision/returnassistant/ReturnAssistantScreenContract.kt
@@ -1,7 +1,6 @@
 package net.gini.android.vision.returnassistant
 
 import android.app.Activity
-import android.content.Intent
 import net.gini.android.vision.GiniVisionBasePresenter
 import net.gini.android.vision.GiniVisionBaseView
 
@@ -18,7 +17,6 @@ interface ReturnAssistantScreenContract {
         fun enablePayButton(selected: Int, total: Int)
         fun disablePayButton(selected: Int, total: Int)
         fun showSelectedLineItemsSum(integralPart: String, fractionPart: String)
-        fun startActivityForResult(intent: Intent, requestCode: Int)
     }
 
     abstract class Presenter(activity: Activity, view: View) :
@@ -27,6 +25,5 @@ interface ReturnAssistantScreenContract {
         abstract fun selectLineItem(lineItem: SelectableLineItem)
         abstract fun deselectLineItem(lineItem: SelectableLineItem)
         abstract fun editLineItem(lineItem: SelectableLineItem)
-        abstract fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?)
     }
 }
